pyproject.toml
[project]
name = "marimo-try"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.45.2",
    "google-genai>=1.1.0",
    "marimo>=0.11.2",
    "openai>=1.61.1",
]
